[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,8 +2,8 @@
 
 int main()
 {
-    int a;
-    int b;
+    long long a;
+    long long b;
     std::cin >> a >> b;
 
     std::cout << (a + b) * (a - b);
